+++ bin/linz-bde-uploader-schema-load.pl
@@ -29,18 +29,20 @@ if ( defined( $ENV{'BDEUPLOADER_SQLDIR'} ) ) {
 our $DB_NAME;
 our $EXTENSION_MODE = 1;
 our $SHOW_VERSION = 0;
+our $READ_ONLY = 0;
 
 sub help
 {
     my ($exitcode) = @_;
-    print STDERR "Usage: $0 [--noextension] { <database> | - }\n";
+    print STDERR "Usage: $0 [--noextension] [--readonly] { <database> | - }\n";
     print STDERR "       $0 --version\n";
     exit $exitcode;
 }
 
 GetOptions (
     "extension!" => \$EXTENSION_MODE,
-    "version!" => \$SHOW_VERSION
+    "version!" => \$SHOW_VERSION,
+    "readonly!" => \$READ_ONLY
 ) || help(0);
 
 $DB_NAME=$ARGV[0];
@@ -167,6 +169,14 @@ foreach my $f (@sqlfiles) {
         print $sql $_;
     }
     close(F);
+}
+
+if ( $READ_ONLY ) {
+    print $sql <<EOF;
+REVOKE UPDATE, INSERT, DELETE, TRUNCATE
+    ON ALL TABLES IN SCHEMA bde_control
+    FROM bde_dba, bde_admin, bde_user;
+EOF
 }
 
 print $sql "COMMIT;\n";